refactor: rename `TemplateCard` to `Card`
Added a component to the `shared/ui` index file.
