--- shared/ui/TemplateCard/index.ts
@@ -1,1 +1,0 @@
-export * from './TemplateCard';
