+++ shared/ui/card/card.module.scss
